cleanup draw code Building
