+++ core/src/com/lipsum/game/entities/Building.java
@@ -1,5 +1,7 @@
 package com.lipsum.game.entities;
 
+import com.badlogic.gdx.graphics.g2d.Batch;
+import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.lipsum.game.LDJam53;
 import com.lipsum.game.event.EventQueue;
 import com.lipsum.game.factory.factories.BuildingFactory;
@@ -46,5 +48,21 @@ public class Building extends Entity{
     @Override
     public void onDispose() {
         remove();
+    }
+
+    private ShapeRenderer renderer = new ShapeRenderer();
+    public void draw (Batch batch, float parentAlpha) {
+        batch.end();
+
+        renderer.setProjectionMatrix(batch.getProjectionMatrix());
+        renderer.setTransformMatrix(batch.getTransformMatrix());
+        renderer.translate(getX(), getY(), 0);
+
+        renderer.begin(ShapeRenderer.ShapeType.Filled);
+        renderer.setColor(getColor());
+        renderer.rect(0, 0, getWidth(), getHeight());
+        renderer.end();
+
+        batch.begin();
     }
 }
